requirements.txt
anyio==3.6.2
click==8.1.3
ecdsa==0.18.0
fastapi==0.85.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
marshmallow==3.18.0
motor==2.5.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.2
pymongo==3.12.3
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
umongo==3.1.0
uvicorn==0.19.0
